refactor: merge benchmark helper functions
