+++ benchmarks/.gitignore
@@ -1,0 +1,2 @@
+*.yaml
+*.yml
